requirement.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
cycler==0.10.0
future==0.18.2
idna==2.7
kiwisolver==1.0.1
lxml==4.5.0
matplotlib==3.0.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.15.2
olefile==0.46
opencv-python==4.2.0.34
pandas==0.23.4
Pillow==7.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytils==0.3
pytz==2018.7
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tools==0.1.9
torch==1.4.0
torchvision==0.5.0
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
